% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_functions.R
\name{b2c_to_hood}
\alias{b2c_to_hood}
\title{Convert a B2C Seurat object to a SpatialExperiment for hoodscanR analysis}
\usage{
b2c_to_hood(obj, annot_col)
}
\arguments{
\item{obj}{A Seurat object (post-aggregated B2C object) with a \code{spatial}
reduction containing SPATIAL_1 and SPATIAL_2 coordinates.}

\item{annot_col}{The name of the metadata column in \code{obj} containing
cell type annotations to pass to \code{hoodscanR::readHoodData}.}
}
\value{
A \code{SpatialExperiment} object ready for hoodscanR analysis.
}
\description{
Convert a B2C Seurat object to a SpatialExperiment for hoodscanR analysis
}
