% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msd_functions.R
\name{color_msd}
\alias{color_msd}
\title{Color MultiSpatialData (MSD) data}
\usage{
color_msd(msd, colors = NULL, intensity = 10)
}
\arguments{
\item{msd}{MSD data frame}

\item{colors}{Color palette}

\item{intensity}{Color intensity}
}
\value{
List of colored MSD data and legend colors
}
\description{
Color MultiSpatialData (MSD) data
}
