% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi_selector.R
\name{crop_b2c}
\alias{crop_b2c}
\title{Crop B2C object to the dimension of selected ROI}
\usage{
crop_b2c(b2c, roi = 1, label.id = "labels_he_expanded")
}
\arguments{
\item{b2c}{B2C object}

\item{roi}{ROI id to crop from}

\item{label.id}{Label ID of the bin2cell segmentation feature}
}
\value{
B2C object
}
\description{
Crop B2C object to the dimension of selected ROI
}
