% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{extract_polygons}
\alias{extract_polygons}
\title{Helper function to extract polygon data from spaceranger object}
\usage{
extract_polygons(obj = NULL, slice = NULL, label.id = "labels_he_expanded")
}
\arguments{
\item{obj}{Seurat object}

\item{slice}{Which slice to use}

\item{label.id}{Label ID}
}
\value{
A data frame with polygon data
}
\description{
Helper function to extract polygon data from spaceranger object
}
