% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_functions.R
\name{get_cor_hood}
\alias{get_cor_hood}
\title{Compute a neighborhood co-localization correlation matrix using hoodscanR}
\usage{
get_cor_hood(spe, k = 100)
}
\arguments{
\item{spe}{A \code{SpatialExperiment} object prepared with \code{b2c_to_hood}.}

\item{k}{Integer. Number of nearest neighbors to use (default 100).}
}
\value{
A correlation matrix of cell-type neighborhood co-localization probabilities.
}
\description{
Compute a neighborhood co-localization correlation matrix using hoodscanR
}
