% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msd_functions.R
\name{get_msd}
\alias{get_msd}
\title{Get MultiSpatialData (MSD) data}
\usage{
get_msd(...)
}
\arguments{
\item{...}{Parameters for FeaturePlot}
}
\value{
MSD data frame
}
\description{
Get MultiSpatialData (MSD) data
}
