% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{load_enact}
\alias{load_enact}
\title{Load ENACT object}
\usage{
load_enact(poly = NULL, post = NULL, path = NULL)
}
\arguments{
\item{poly}{ENACT table (csv) containing polygon data}

\item{post}{Post-aggregated (with ENACT) Seurat object}

\item{path}{Path to the image file used for ENACT segmentation}
}
\value{
B2C object (A list containing pre, post, and image data, adapted from ENACT objects)
}
\description{
Load ENACT object
}
