% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{load_img}
\alias{load_img}
\title{Load image file}
\usage{
load_img(path)
}
\arguments{
\item{path}{Path to the image file used for bin2cell segmentation}
}
\value{
A data frame with image data
}
\description{
Load image file
}
