% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{overview_b2c}
\alias{overview_b2c}
\title{Plot spatial feature from a B2C object (requires scaled-down image)}
\usage{
overview_b2c(
  b2c,
  feat,
  pt.size = 0.001,
  he_alpha = 0.4,
  col.low = "gray",
  col.high = "seagreen2",
  scalebar = 500
)
}
\arguments{
\item{b2c}{B2C object}

\item{feat}{Gene feature to plot}

\item{pt.size}{Size of the points}

\item{he_alpha}{Alpha value for H&E image}

\item{col.low}{Low color for colorscale}

\item{col.high}{High color for colorscale}

\item{scalebar}{scalebar size in microns, FALSE for no scalebar}
}
\description{
Plot spatial feature from a B2C object (requires scaled-down image)
}
