% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_b2c_nn}
\alias{plot_b2c_nn}
\title{Plot B2C object using nearest neighbor}
\usage{
plot_b2c_nn(
  b2c = NULL,
  features = NULL,
  roi = 1,
  colors = NULL,
  intensity = 10,
  he_alpha = 0.4,
  pt_size = 0.5,
  plot.type = c("points", "hulls"),
  label.id = "labels_he_expanded",
  outline.hulls = NULL,
  show.labels = F
)
}
\arguments{
\item{b2c}{B2C object}

\item{features}{Features to plot}

\item{roi}{index of the ROI to plot}

\item{colors}{Color palette}

\item{intensity}{Color intensity}

\item{he_alpha}{Alpha value for H&E image}

\item{pt_size}{Point size}

\item{plot.type}{Type of plot}

\item{label.id}{Label ID of the bin2cell segmentation feature}

\item{outline.hulls}{character vector of hulls to outline (by expanded labels ID)}

\item{show.labels}{Whether or not to plot the hulls labels}
}
\description{
Plot B2C object using nearest neighbor
}
