% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_dist}
\alias{plot_dist}
\title{Plot cell-cell distances distribution from a ROI plot (previously quantified with get_dist)}
\usage{
plot_dist(df, binwidth = 5)
}
\arguments{
\item{df}{Data frame of distances (output from get_dist)}

\item{binwidth}{Binwidth of the histograms}
}
\description{
Plot cell-cell distances distribution from a ROI plot (previously quantified with get_dist)
}
