% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_msd}
\alias{plot_msd}
\title{Plot MSD data}
\usage{
plot_msd(
  object,
  features = NULL,
  img,
  msd,
  coord,
  he_alpha = 0.4,
  pt_size = 1,
  plot = TRUE
)
}
\arguments{
\item{object}{Post-segemented Seurat object}

\item{features}{Features to plot}

\item{img}{Image data}

\item{msd}{Mulit Spatial Data}

\item{coord}{Coordinates}

\item{he_alpha}{Alpha value for H&E image}

\item{pt_size}{Point size}

\item{plot}{Whether to plot or return ggplot object}
}
\value{
ggplot object or plot
}
\description{
Plot MSD data
}
