% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_segmentation}
\alias{plot_segmentation}
\title{Plot Segmentation from a B2C (cropped) object}
\usage{
plot_segmentation(
  b2c,
  label.id = "labels_he_expanded",
  he_alpha = 1,
  color = "black",
  linewidth = 0.1
)
}
\arguments{
\item{b2c}{B2C (cropped) object}

\item{label.id}{Label ID of the bin2cell segmentation feature}

\item{he_alpha}{Alpha value for H&E image}

\item{color}{Outline color for the polygons}

\item{linewidth}{Linewidth for the polygons}
}
\description{
Plot Segmentation from a B2C (cropped) object
}
