% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi_selector.R
\name{roi_selector}
\alias{roi_selector}
\title{ROI Selector}
\usage{
roi_selector(path, b2c = NULL)
}
\arguments{
\item{path}{Path to the image file used for bin2cell segmentation}

\item{b2c}{B2C object}
}
\value{
list of ROI coordinates
}
\description{
ROI Selector
}
