% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{scaledown_img}
\alias{scaledown_img}
\title{Scale down original image}
\usage{
scaledown_img(b2c, grid.size = 10)
}
\arguments{
\item{b2c}{B2C object}

\item{grid.size}{grid size defining the resolution of the scaled-down image}
}
\value{
B2C object
}
\description{
Scale down original image
}
