% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{simpleplot}
\alias{simpleplot}
\title{Plot spatial feature from a B2C object (requires scaled-down image)}
\usage{
simpleplot(b2c, feature, pt.size = 0.001, he_alpha = 0.4)
}
\arguments{
\item{b2c}{B2C object}

\item{feature}{Gene feature to plot}

\item{pt.size}{Size of the points}
}
\description{
Plot spatial feature from a B2C object (requires scaled-down image)
}
