% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{space2b2c}
\alias{space2b2c}
\title{Helper function to convert spaceranger to bin2cell format}
\usage{
space2b2c(obj = NULL, slice = NULL)
}
\arguments{
\item{obj}{Spaceranger object to convert}

\item{slice}{Which slice to use}
}
\value{
A bin2cell-like object
}
\description{
Helper function to convert spaceranger to bin2cell format
}
