% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{upscale_roi}
\alias{upscale_roi}
\title{Upscale ROI image using original OME-TIFF image}
\usage{
upscale_roi(b2c, path, series = 1, resolution = 1)
}
\arguments{
\item{b2c}{B2C (cropped) object}

\item{path}{Path to the original OME-TIFF image}

\item{series}{Index of the selected OME-TIFF image series}

\item{resolution}{Index of the selected OME-TIFF image resolution}
}
\value{
B2C (cropped) object
}
\description{
Upscale ROI image using original OME-TIFF image
}
