% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{wrap_msd}
\alias{wrap_msd}
\title{Wrapper for MSD plotting}
\usage{
wrap_msd(
  object,
  img,
  features,
  coord,
  intensity = 10,
  he_alpha = 0.4,
  pt_size = 0.5,
  plot = TRUE
)
}
\arguments{
\item{object}{Post-aggregated Seurat object}

\item{img}{Image data}

\item{features}{Features to plot}

\item{coord}{Coordinates}

\item{intensity}{Color intensity}

\item{he_alpha}{Alpha value for H&E image}

\item{pt_size}{Point size}

\item{plot}{Whether to plot or return ggplot object}
}
\value{
ggplot object or plot
}
\description{
Wrapper for MSD plotting
}
