% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_functions.R
\name{get_dist}
\alias{get_dist}
\title{Function to calculate neighbors within a specified radius}
\usage{
get_dist(plot, radius = 100)
}
\arguments{
\item{plot}{A ROI feature plot generated by plot_b2c}

\item{radius}{The maximum distance (microns) within which neighbors are to be detected}
}
\value{
A data frame with distances data
}
\description{
Function to calculate neighbors within a specified radius
}
