% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{load_b2c}
\alias{load_b2c}
\title{Load B2C object}
\usage{
load_b2c(pre = NULL, post = NULL, path = NULL)
}
\arguments{
\item{pre}{Pre-segemented Seurat object}

\item{post}{Post-segemented Seurat object}

\item{path}{Path to the image file used for bin2cell segmentation}
}
\value{
B2C object (A list containing pre, post, and image data)
}
\description{
Load B2C object
}
