% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{overview_b2c}
\alias{overview_b2c}
\title{Plot spatial feature from a B2C object (requires scaled-down image)}
\usage{
overview_b2c(
  b2c,
  feat,
  pt.size = 0.001,
  he_alpha = 0.4,
  col.low = "gray",
  col.high = "seagreen2"
)
}
\arguments{
\item{b2c}{B2C object}

\item{feat}{Gene feature to plot}

\item{pt.size}{Size of the points}
}
\description{
Plot spatial feature from a B2C object (requires scaled-down image)
}
