% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_b2c}
\alias{plot_b2c}
\title{Plot B2C object}
\usage{
plot_b2c(
  b2c = NULL,
  features = NULL,
  colors = NULL,
  intensity = 10,
  he_alpha = 0.4,
  pt_size = 0.5,
  plot.type = c("points", "hulls"),
  label.id = "labels_he_expanded"
)
}
\arguments{
\item{b2c}{B2C object}

\item{features}{Features to plot}

\item{colors}{Color palette}

\item{intensity}{Color intensity}

\item{he_alpha}{Alpha value for H&E image}

\item{pt_size}{Point size}

\item{plot.type}{Type of plot}

\item{label.id}{Label ID of the bin2cell segmentation feature}
}
\description{
Plot B2C object
}
