% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_b2c}
\alias{plot_b2c}
\title{Plot B2C object}
\usage{
plot_b2c(
  b2c,
  feat,
  label.id = "labels_he_expanded",
  min.visible = 0,
  col.low = NULL,
  col.mid = NULL,
  col.high = "orangered",
  alpha.low = 0,
  alpha.mid = 0.5,
  alpha.high = 1,
  scale.min.max = NULL,
  pt.size = 1,
  shape = NULL,
  he_alpha = 0.3,
  title = NULL,
  plot.type = c("points", "hulls"),
  show.bins = "no",
  outline.hulls = NULL,
  show.labels = F,
  plot = T,
  scalebar = 200,
  scalebar.width = 10,
  translate = T,
  filter.feat = NULL,
  filter.threshold = 0
)
}
\arguments{
\item{b2c}{B2C object}

\item{feat}{Features to plot}

\item{label.id}{Label ID of the bin2cell segmentation feature}

\item{min.visible}{Feature value threshold for visualization (single value or vector for each feature)}

\item{col.low}{Low color for colorscale (single value or vector for each feature)}

\item{col.mid}{Mid color for colorscale (single value or vector for each feature)}

\item{col.high}{High color for colorscale (single value or vector for each feature)}

\item{alpha.low}{Low alpha value for colorscale (single value or vector for each feature)}

\item{alpha.mid}{Mid alpha value for colorscale (single value or vector for each feature)}

\item{alpha.high}{High alpha value for colorscale (single value or vector for each feature)}

\item{scale.min.max}{List of vectors (of length 2) indicating the min and max value for the color gradient scale}

\item{shape}{Shape of the points. Defaults to solid circle (16) for 1 feature, empty circle (21) for multi-feature plot to better see multi-positive cells}

\item{he_alpha}{Alpha value for H&E image}

\item{title}{Plot title}

\item{plot.type}{Type of plot (points or hulls or both)}

\item{show.bins}{Whether to show the 2um bins data. Defaults to "no", use "yes" to show bins corresponding to the feature and "all" to show all bins data (in white)}

\item{outline.hulls}{Character vector of hulls to outline (by expanded labels ID)}

\item{show.labels}{Whether or not to plot the hulls labels}

\item{plot}{Whether to display or return the plot and the data}

\item{scalebar}{Scalebar size in microns, FALSE for no scalebar}

\item{scalebar.width}{width of the scalebar}

\item{translate}{Whether or not to translate the plot to the (0, 0) origin (can be useful to adjust plot size when comparing multiple ROIs)}

\item{filter.feat}{Features (list of vectors) to pre-filter the data (e.g. keep only cells expressing these features, order has to match order used in feat, use "" for no filtering)}

\item{filter.threshold}{Threshold (list of vectors) levels for filter.feat}

\item{pt_size}{Point size}
}
\description{
Plot B2C object
}
